requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.2
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
Flask==3.0.2
google-auth==2.28.0
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
pillow==10.2.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.26
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
django-extensions==3.2.3